fix: fix overlay auto-signIn if no session
